# docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    email: support@karimastore.com
    name: Karima Store API Support
  description: |-
    Karima Store E-commerce API with JWT Authentication

    ## Authentication

    This API uses **JWT (JSON Web Token)** for authentication.

    ### For Web/Browser Clients:
    1. Login via /api/v1/auth/login endpoint
    2. Session cookie (jwt) will be set automatically
    3. Make API requests with the cookie included

    ### For API/Mobile Clients:
    1. Obtain JWT token from /api/v1/auth/login endpoint
    2. Include token in requests:
    - Method 1: Authorization: Bearer <jwt_token>
    - Method 2: X-Session-Token: <jwt_token> header

    ### Authorization Levels:
    - **Public**: No authentication required (GET endpoints for browsing)
    - **Authenticated**: Valid JWT token required
    - **Admin**: Valid JWT token + admin role
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Karima Store API
  version: "1.0"
paths:
...
  /api/v1/whatsapp/order-created/{order_id}:
    get:
      consumes:
      - application/json
      description: 'Send notification when order is created. **Admin only**: Requires
        authentication with admin role.'
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: No valid session or session expired'
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 'Forbidden: Insufficient permissions (admin role required)'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Order not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - 'BearerAuth ':
        - ""
      summary: Send order created notification (Admin only)
      tags:
      - whatsapp
  /api/v1/whatsapp/payment-success/{order_id}:
    get:
      consumes:
      - application/json
      description: 'Send notification when payment is successful. **Admin only**:
        Requires authentication with admin role.'
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: No valid session or session expired'
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 'Forbidden: Insufficient permissions (admin role required)'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Order not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - 'BearerAuth ':
        - ""
      summary: Send payment success notification (Admin only)
      tags:
      - whatsapp
  /api/v1/whatsapp/send:
    post:
      consumes:
      - application/json
      description: 'Send a message to a WhatsApp number. **Admin only**: Requires
        authentication with admin role.'
      parameters:
      - description: Phone number in E.164 format
        in: query
        name: phoneNumber
        required: true
        type: string
      - description: Message content
        in: body
        name: message
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: No valid session or session expired'
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 'Forbidden: Insufficient permissions (admin role required)'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - 'BearerAuth ':
        - ""
      summary: Send WhatsApp message (Admin only)
      tags:
      - whatsapp
  /api/v1/whatsapp/status:
    get:
      consumes:
      - application/json
      description: Get current status of WhatsApp service (public endpoint, no authentication
        required)
      produces:
      - application/json
      responses:
        "200":
          description: Status response
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get WhatsApp status
      tags:
      - whatsapp
  /api/v1/whatsapp/test:
    post:
      consumes:
      - application/json
      description: 'Send a test message to verify WhatsApp integration. **Admin only**:
        Requires authentication with admin role.'
      parameters:
      - description: Phone number in E.164 format
        in: query
        name: phoneNumber
        required: true
        type: string
      - description: Message content
        in: body
        name: message
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: No valid session or session expired'
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 'Forbidden: Insufficient permissions (admin role required)'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - 'BearerAuth ':
        - ""
      summary: Send test WhatsApp message (Admin only)
      tags:
      - whatsapp
  /api/v1/whatsapp/webhook:
    post:
      consumes:
      - application/json
      description: Handle incoming webhook events from WhatsApp (public endpoint,
        no authentication required)
      parameters:
      - description: Webhook data
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Process WhatsApp webhook
      tags:
      - whatsapp
  /api/v1/whatsapp/webhook-url:
    get:
      consumes:
      - application/json
      description: Get the webhook URL for WhatsApp
      produces:
      - application/json
      responses:
        "200":
          description: Webhook URL response
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get WhatsApp webhook URL
      tags:
      - whatsapp
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization:
      Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"